# [FLINK-21736][state] Introduce state scope latency tracking metrics