--- a/flink-core/src/main/java/org/apache/flink/configuration/CheckpointingOptions.java
+++ b/flink-core/src/main/java/org/apache/flink/configuration/CheckpointingOptions.java
@@ -26,7 +26,7 @@ import org.apache.flink.configuration.description.TextElement;
 public class CheckpointingOptions {
 
     // ------------------------------------------------------------------------
-    //  general checkpoint and state backend options
+    //  general checkpoint options
     // ------------------------------------------------------------------------
 
     /**
@@ -39,8 +39,12 @@ public class CheckpointingOptions {
      * StateBackendFactory#createFromConfig(ReadableConfig, ClassLoader)} method is called.
      *
      * <p>Recognized shortcut names are 'hashmap' and 'rocksdb'.
+     *
+     * @deprecated Use {@link StateBackendOptions#STATE_BACKEND}.
      */
-    @Documentation.Section(value = Documentation.Sections.COMMON_STATE_BACKENDS, position = 1)
+    @Documentation.Section(value = Documentation.Sections.COMMON_STATE_BACKENDS)
+    @Documentation.ExcludeFromDocumentation("Hidden for deprecated")
+    @Deprecated
     public static final ConfigOption<String> STATE_BACKEND =
             ConfigOptions.key("state.backend")
                     .stringType()