--- a/flink-core/src/main/java/org/apache/flink/configuration/StateBackendOptions.java
+++ b/flink-core/src/main/java/org/apache/flink/configuration/StateBackendOptions.java
@@ -61,4 +61,31 @@ public class StateBackendOptions {
                                     .linebreak()
                                     .text("Recognized shortcut names are 'hashmap' and 'rocksdb'.")
                                     .build());
+
+    @Documentation.Section(Documentation.Sections.STATE_BACKEND_LATENCY_TRACKING)
+    public static final ConfigOption<Boolean> LATENCY_TRACK_ENABLED =
+            ConfigOptions.key("state.backend.latency-track.keyed-state-enabled")
+                    .booleanType()
+                    .defaultValue(false)
+                    .withDescription(
+                            "Whether to track latency of keyed state operations, e.g value state put/get/clear.");
+
+    @Documentation.Section(Documentation.Sections.STATE_BACKEND_LATENCY_TRACKING)
+    public static final ConfigOption<Integer> LATENCY_TRACK_SAMPLE_INTERVAL =
+            ConfigOptions.key("state.backend.latency-track.sample-interval")
+                    .intType()
+                    .defaultValue(100)
+                    .withDescription(
+                            String.format(
+                                    "The sample interval of latency track once '%s' is enabled. "
+                                            + "The default value is 100, which means we would track the latency every 100 access requests.",
+                                    LATENCY_TRACK_ENABLED.key()));
+
+    @Documentation.Section(Documentation.Sections.STATE_BACKEND_LATENCY_TRACKING)
+    public static final ConfigOption<Integer> LATENCY_TRACK_HISTORY_SIZE =
+            ConfigOptions.key("state.backend.latency-track.history-size")
+                    .intType()
+                    .defaultValue(128)
+                    .withDescription(
+                            "Defines the number of measured latencies to maintain at each state access operation.");
 }