--- a/flink-runtime/src/main/java/org/apache/flink/runtime/state/AbstractKeyedStateBackendBuilder.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/state/AbstractKeyedStateBackendBuilder.java
@@ -22,6 +22,7 @@ import org.apache.flink.api.common.ExecutionConfig;
 import org.apache.flink.api.common.typeutils.TypeSerializer;
 import org.apache.flink.core.fs.CloseableRegistry;
 import org.apache.flink.runtime.query.TaskKvStateRegistry;
+import org.apache.flink.runtime.state.metrics.LatencyTrackingStateConfig;
 import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
 
 import org.slf4j.Logger;
@@ -43,6 +44,7 @@ public abstract class AbstractKeyedStateBackendBuilder<K>
     protected final KeyGroupRange keyGroupRange;
     protected final ExecutionConfig executionConfig;
     protected final TtlTimeProvider ttlTimeProvider;
+    protected final LatencyTrackingStateConfig latencyTrackingStateConfig;
     protected final StreamCompressionDecorator keyGroupCompressionDecorator;
     protected final Collection<KeyedStateHandle> restoreStateHandles;
     protected final CloseableRegistry cancelStreamRegistry;
@@ -55,6 +57,7 @@ public abstract class AbstractKeyedStateBackendBuilder<K>
             KeyGroupRange keyGroupRange,
             ExecutionConfig executionConfig,
             TtlTimeProvider ttlTimeProvider,
+            LatencyTrackingStateConfig latencyTrackingStateConfig,
             @Nonnull Collection<KeyedStateHandle> stateHandles,
             StreamCompressionDecorator keyGroupCompressionDecorator,
             CloseableRegistry cancelStreamRegistry) {
@@ -66,6 +69,7 @@ public abstract class AbstractKeyedStateBackendBuilder<K>
         this.keyGroupRange = keyGroupRange;
         this.executionConfig = executionConfig;
         this.ttlTimeProvider = ttlTimeProvider;
+        this.latencyTrackingStateConfig = latencyTrackingStateConfig;
         this.keyGroupCompressionDecorator = keyGroupCompressionDecorator;
         this.restoreStateHandles = stateHandles;
         this.cancelStreamRegistry = cancelStreamRegistry;