--- a/flink-runtime/src/main/java/org/apache/flink/runtime/state/StateBackendLoader.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/state/StateBackendLoader.java
@@ -22,6 +22,7 @@ import org.apache.flink.configuration.CheckpointingOptions;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.IllegalConfigurationException;
 import org.apache.flink.configuration.ReadableConfig;
+import org.apache.flink.configuration.StateBackendOptions;
 import org.apache.flink.runtime.state.delegate.DelegatingStateBackend;
 import org.apache.flink.runtime.state.hashmap.HashMapStateBackend;
 import org.apache.flink.runtime.state.hashmap.HashMapStateBackendFactory;
@@ -79,7 +80,7 @@ public class StateBackendLoader {
 
     /**
      * Loads the unwrapped state backend from the configuration, from the parameter 'state.backend',
-     * as defined in {@link CheckpointingOptions#STATE_BACKEND}.
+     * as defined in {@link StateBackendOptions#STATE_BACKEND}.
      *
      * <p>The state backends can be specified either via their shortcut name, or via the class name
      * of a {@link StateBackendFactory}. If a StateBackendFactory class name is specified, the
@@ -109,7 +110,7 @@ public class StateBackendLoader {
         checkNotNull(config, "config");
         checkNotNull(classLoader, "classLoader");
 
-        final String backendName = config.get(CheckpointingOptions.STATE_BACKEND);
+        final String backendName = config.get(StateBackendOptions.STATE_BACKEND);
         if (backendName == null) {
             return null;
         }
@@ -175,7 +176,7 @@ public class StateBackendLoader {
                 } catch (ClassCastException | InstantiationException | IllegalAccessException e) {
                     throw new DynamicCodeLoadingException(
                             "The class configured under '"
-                                    + CheckpointingOptions.STATE_BACKEND.key()
+                                    + StateBackendOptions.STATE_BACKEND.key()
                                     + "' is not a valid state backend factory ("
                                     + backendName
                                     + ')',