--- a/flink-runtime/src/main/java/org/apache/flink/runtime/state/hashmap/HashMapStateBackend.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/state/hashmap/HashMapStateBackend.java
@@ -40,6 +40,7 @@ import org.apache.flink.runtime.state.OperatorStateHandle;
 import org.apache.flink.runtime.state.TaskStateManager;
 import org.apache.flink.runtime.state.heap.HeapKeyedStateBackendBuilder;
 import org.apache.flink.runtime.state.heap.HeapPriorityQueueSetFactory;
+import org.apache.flink.runtime.state.metrics.LatencyTrackingStateConfig;
 import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
 
 import javax.annotation.Nonnull;
@@ -76,10 +77,18 @@ public class HashMapStateBackend extends AbstractStateBackend implements Configu
 
     // -----------------------------------------------------------------------
 
+    /** Creates a new state backend. */
+    public HashMapStateBackend() {}
+
+    private HashMapStateBackend(HashMapStateBackend original, ReadableConfig config) {
+        // configure latency tracking
+        latencyTrackingConfigBuilder = original.latencyTrackingConfigBuilder.configure(config);
+    }
+
     @Override
     public HashMapStateBackend configure(ReadableConfig config, ClassLoader classLoader)
             throws IllegalConfigurationException {
-        return this;
+        return new HashMapStateBackend(this, config);
     }
 
     @Override
@@ -102,6 +111,8 @@ public class HashMapStateBackend extends AbstractStateBackend implements Configu
         HeapPriorityQueueSetFactory priorityQueueSetFactory =
                 new HeapPriorityQueueSetFactory(keyGroupRange, numberOfKeyGroups, 128);
 
+        LatencyTrackingStateConfig latencyTrackingStateConfig =
+                latencyTrackingConfigBuilder.setMetricGroup(metricGroup).build();
         return new HeapKeyedStateBackendBuilder<>(
                         kvStateRegistry,
                         keySerializer,
@@ -110,6 +121,7 @@ public class HashMapStateBackend extends AbstractStateBackend implements Configu
                         keyGroupRange,
                         env.getExecutionConfig(),
                         ttlTimeProvider,
+                        latencyTrackingStateConfig,
                         stateHandles,
                         getCompressionDecorator(env.getExecutionConfig()),
                         localRecoveryConfig,