--- a/flink-runtime/src/main/java/org/apache/flink/runtime/state/memory/MemoryStateBackend.java
+++ b/flink-runtime/src/main/java/org/apache/flink/runtime/state/memory/MemoryStateBackend.java
@@ -41,6 +41,7 @@ import org.apache.flink.runtime.state.TaskStateManager;
 import org.apache.flink.runtime.state.filesystem.AbstractFileStateBackend;
 import org.apache.flink.runtime.state.heap.HeapKeyedStateBackendBuilder;
 import org.apache.flink.runtime.state.heap.HeapPriorityQueueSetFactory;
+import org.apache.flink.runtime.state.metrics.LatencyTrackingStateConfig;
 import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
 import org.apache.flink.util.TernaryBoolean;
 
@@ -257,6 +258,10 @@ public class MemoryStateBackend extends AbstractFileStateBackend
         super(original.getCheckpointPath(), original.getSavepointPath(), configuration);
 
         this.maxStateSize = original.maxStateSize;
+
+        // configure latency tracking
+        latencyTrackingConfigBuilder =
+                original.latencyTrackingConfigBuilder.configure(configuration);
     }
 
     // ------------------------------------------------------------------------
@@ -347,6 +352,8 @@ public class MemoryStateBackend extends AbstractFileStateBackend
         TaskStateManager taskStateManager = env.getTaskStateManager();
         HeapPriorityQueueSetFactory priorityQueueSetFactory =
                 new HeapPriorityQueueSetFactory(keyGroupRange, numberOfKeyGroups, 128);
+        LatencyTrackingStateConfig latencyTrackingStateConfig =
+                latencyTrackingConfigBuilder.setMetricGroup(metricGroup).build();
         return new HeapKeyedStateBackendBuilder<>(
                         kvStateRegistry,
                         keySerializer,
@@ -355,6 +362,7 @@ public class MemoryStateBackend extends AbstractFileStateBackend
                         keyGroupRange,
                         env.getExecutionConfig(),
                         ttlTimeProvider,
+                        latencyTrackingStateConfig,
                         stateHandles,
                         AbstractStateBackend.getCompressionDecorator(env.getExecutionConfig()),
                         taskStateManager.createLocalRecoveryConfig(),