## What is the purpose of the change

Track the state latency metrics.

## Brief change log

1. Separate state backend and checkpoint options.
2. Introduce latency tracking state config and basic classes.
3. Enable to tack latency state with introduced `LatencyTrackingValueState`, `LatencyTrackingListState`, `LatencyTrackingMapState`, `LatencyTrackingAggregatingState` and `LatencyTrackingReducingState`.


## Verifying this change

Added test `LatencyTrackingStateConfigTest` to verify configuration test.

Added test `LatencyTrackingValueStateTest`, `LatencyTrackingListStateTest`, `LatencyTrackingMapStateTest`, `LatencyTrackingAggregatingStateTest` and `LatencyTrackingReducingStateTest` to verify different state wrapper.

## Does this pull request potentially affect one of the following parts:

  - Dependencies (does it add or upgrade a dependency): yes (add dependency of `io.dropwizard.metrics:metrics-core:3.2.6
`)
  - The public API, i.e., is any changed class annotated with `@Public(Evolving)`: no
  - The serializers: no
  - The runtime per-record code paths (performance sensitive): yes, state with metrics would impact the state performance.
  - Anything that affects deployment or recovery: JobManager (and its components), Checkpointing, Kubernetes/Yarn/Mesos, ZooKeeper: no
  - The S3 file system connector: no

## Documentation

  - Does this pull request introduce a new feature? yes
  - If yes, how is the feature documented? docs