@Myasuka 2021-03-16
If we enable track the latency of state access, there would exist some performance regression as we need to calculate the latency via two times call of `System.nanoTime`.
Below is the performance compare with enabled latency tracking state (sample interval: 100). We can see heap state backend would be impacted more obviously.

operations | type | latency tracking QPS | original QPS | performance change
-- | -- | -- | -- | --
ListStateBenchmark.listAdd | HEAP | 8140.139 | 9064.081 | -10.19%
ListStateBenchmark.listAddAll | HEAP | 1170.439 | 1086.373 | 7.74%
ListStateBenchmark.listAppend | HEAP | 4070.654 | 4364.116 | -6.72%
ListStateBenchmark.listGet | HEAP | 4112.02 | 4403.857 | -6.63%
ListStateBenchmark.listGetAndIterate | HEAP | 4028.173 | 4293.069 | -6.17%
ListStateBenchmark.listUpdate | HEAP | 3977.07 | 4302.616 | -7.57%
MapStateBenchmark.mapAdd | HEAP | 4272.695 | 4429.009 | -3.53%
MapStateBenchmark.mapContains | HEAP | 3606.731 | 3727.85 | -3.25%
MapStateBenchmark.mapEntries | HEAP | 19798.66 | 22276.277 | -11.12%
MapStateBenchmark.mapGet | HEAP | 3430.402 | 3505.954 | -2.15%
MapStateBenchmark.mapIsEmpty | HEAP | 4833.259 | 4981.474 | -2.98%
MapStateBenchmark.mapIterator | HEAP | 20472.585 | 21335.119 | -4.04%
MapStateBenchmark.mapKeys | HEAP | 22864.626 | 25057.534 | -8.75%
MapStateBenchmark.mapPutAll | HEAP | 2285.514 | 2409.35 | -5.14%
MapStateBenchmark.mapRemove | HEAP | 4487.178 | 4717.213 | -4.88%
MapStateBenchmark.mapUpdate | HEAP | 3443.089 | 3555.822 | -3.17%
MapStateBenchmark.mapValues | HEAP | 23053.149 | 25037.289 | -7.92%
ValueStateBenchmark.valueAdd | HEAP | 4869.279 | 5026.357 | -3.13%
ValueStateBenchmark.valueGet | HEAP | 4880.446 | 5030.63 | -2.99%
ValueStateBenchmark.valueUpdate | HEAP | 4878.84 | 5030.789 | -3.02%
ListStateBenchmark.listAdd | ROCKSDB | 661.859 | 676.74 | -2.20%
ListStateBenchmark.listAddAll | ROCKSDB | 387.586 | 392.964 | -1.37%
ListStateBenchmark.listAppend | ROCKSDB | 651.417 | 659.463 | -1.22%
ListStateBenchmark.listGet | ROCKSDB | 196.54 | 196.321 | 0.11%
ListStateBenchmark.listGetAndIterate | ROCKSDB | 196.15 | 197.117 | -0.49%
ListStateBenchmark.listUpdate | ROCKSDB | 655.848 | 668.958 | -1.96%
MapStateBenchmark.mapAdd | ROCKSDB | 564.704 | 570.076 | -0.94%
MapStateBenchmark.mapContains | ROCKSDB | 72.266 | 72.527 | -0.36%
MapStateBenchmark.mapEntries | ROCKSDB | 497.074 | 505.535 | -1.67%
MapStateBenchmark.mapGet | ROCKSDB | 72.397 | 72.257 | 0.19%
MapStateBenchmark.mapIsEmpty | ROCKSDB | 65.11 | 64.307 | 1.25%
MapStateBenchmark.mapIterator | ROCKSDB | 499.665 | 502.106 | -0.49%
MapStateBenchmark.mapKeys | ROCKSDB | 501.688 | 507.647 | -1.17%
MapStateBenchmark.mapPutAll | ROCKSDB | 162.493 | 166.482 | -2.40%
MapStateBenchmark.mapRemove | ROCKSDB | 588.882 | 592.891 | -0.68%
MapStateBenchmark.mapUpdate | ROCKSDB | 564.878 | 565.93 | -0.19%
MapStateBenchmark.mapValues | ROCKSDB | 504.755 | 504.891 | -0.03%
ValueStateBenchmark.valueAdd | ROCKSDB | 565.801 | 566.092 | -0.05%
ValueStateBenchmark.valueGet | ROCKSDB | 924.21 | 929.503 | -0.57%
ValueStateBenchmark.valueUpdate | ROCKSDB | 558.719 | 576.711 | -3.12%