@Myasuka 2021-03-25
Unfortunately, changing the histogram implementation from `com.codahale.metrics.Histogram + SlidingTimeWindowReservoir` to `DescriptiveStatisticsHistogram + CircularDoubleArray` gets a worse performance regression via flink-benchmarks, still investigating ..